Разбивает игру на модули
