+++ utils.js
@@ -1,0 +1,3 @@
+const randomNumbers = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+
+export default randomNumbers;
